refactor: 도메인 응집
