+++ src/entities/project/ui/ProjectAccordion.tsx
