+++ src/entities/trouble/mappers/communityCard.mapper.ts
