+++ src/entities/trouble/mappers/communityPostDetail.mapper.ts
