+++ src/entities/trouble/mappers/myPostDetail.mapper.ts
