+++ src/features/mypage/ui/WithdrawBox.tsx
