+++ src/hooks/useCommunityCards.ts
@@ -4,7 +4,7 @@ import type {
   CommunityServerCard,
   CommunitySort,
 } from "@/types/community.model";
-import { toCommunityCards } from "@/mappers/communityCard.mapper";
+import { toCommunityCards } from "@/entities/trouble/mappers/communityCard.mapper";
 import type { TroublogCardProps } from "@/entities/trouble/ui/TroublogCard";
 
 type PageResp = {
