+++ src/pages/Home/HomePage.tsx
@@ -2,8 +2,8 @@ import { useState, useCallback, useEffect, useRef } from "react";
 import PostButton from "@/components/Button/PostButton";
 import Snackbar from "@/components/Feedback/Snackbar";
 import TroublogCard from "@/entities/trouble/ui/TroublogCard";
-import ProjectAccordion from "@/components/Project/ProjectAccordion";
-import ProjectFolderCard from "@/components/Project/ProjectFolderCard";
+import ProjectAccordion from "@/entities/project/ui/ProjectAccordion";
+import ProjectFolderCard from "@/entities/project/ui/ProjectFolderCard";
 import FolderModal from "@/components/Modal/FolderModal";
 import useClickOutside from "@/hooks/useClickOutside";
 import { getProjectList, postCreateProject } from "@/api/project.api";
